% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carSignal-v15.9.10.R
\name{get.best.templates}
\alias{get.best.templates}
\title{Busca la curva que mejor se ajuste a los templates entregados como entrada, con el propósito de calcular
el índice ARI.}
\usage{
get.best.templates(time.instants, signal, templates,
  referential.time.instant = 0, delta.time.before.ref = 0,
  delta.time.after.ref = 20 * 0.8, comparison.function = get.MSE,
  keep.details = TRUE, time.tol = min(diff(time.instants))/100, ...)
}
\arguments{
\item{time.instants:}{lista con los tiempos de muestreo de la señal}

\item{signal:}{señal con la que se desea comparar los templates}

\item{templates:}{templates a comparar con la señal}

\item{comparison.function:}{función con la cual se evalúa la curva con mejor ajuste a los template.}
}
\value{
El mejor valor de ajuste al template y una lista con los valores del template.
}
\description{
Busca la curva que mejor se ajuste a los templates entregados como entrada, con el propósito de calcular
el índice ARI.
}
\references{
Frank P. Tiecks, Arthur M. Lam, Rune Aaslid, David W. Newell. Comparison of Static and Dynamic Cerebral Autoregulation Measurements.
Stroke, 1995, 26, pages 1014-1019.
}

