% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carSignal-v15.9.10.R
\name{normalise.signal}
\alias{normalise.signal}
\title{Normalización de una señal}
\usage{
normalise.signal(signal, signal.baseline.value, signal.min.value)
}
\arguments{
\item{signal:}{La señal que se desea normalizar}

\item{signal.baseline.value:}{valor basal entorno al cual se desea normalizar}

\item{signal.min.value:}{El valor mínimo que tiene la señal}
}
\value{
La señal normalizada
}
\description{
Normalización de una señal
}

